requirements.txt
instaloader
requests
schedule
google-generativeai
firebase-admin
python-dotenv
google-search-results
